Drive the reduce-tier shared-row staging off a scheduler-stamped Stage
The fused norm→linear shared-row prologue was the one staging decision
still detected at materialize time. Move the detection scheduler-side:
_schedule._row_stage runs when a cooperative reduce partition is chosen
and stamps a depth-1 sync Stage whose smem names the reused input row (a
derived schedule field, never a knob — tile_signature / featurization
untouched); the unmapped TileOp is seeded with the recognized LoopOp's
inputs so the scheduler can read operand shapes. _bind_reduce now only
APPLIES the stage off ctx.stage (sync_row_fill + _restage_loads), the
same Stage → apply shape the contraction tiers follow; the detection
helpers are deleted from _factor.py.

Bit-identical (fused RMSNorm→Linear x_smem, softmax, mean dumps diff
clean); full suite green.
